tools/simple-test/requirements.txt
attrs==19.3.0
importlib-metadata==1.7.0
jsonschema==3.2.0
pyrsistent==0.16.0
PyYAML==5.3.1
six==1.15.0
zipp==3.1.0
